# specialize matrix multiplication with Diagonals