Fixes #31325. This is a continuation of #31400, which got stalled. Closes #31400.

The following benchmarks tell pretty much everything about which specialized methods were missing and hence falling back to generic matmuls.

```julia
# before the PR
julia> using LinearAlgebra, BenchmarkTools, Test

julia> A = rand(100, 100);

julia> B = similar(A);

julia> D = Diagonal(rand(100));

julia> for transformA in (identity, adjoint, transpose), transformD in (identity, Adjoint, Transpose)
           @btime mul!($(copy(B)), $(transformA(A)), $(transformD(D)))
           @btime mul!($(copy(B)), $(transformD(D)), $(transformA(A)))
       end
  791.282 μs (6 allocations: 336 bytes)
  2.169 μs (0 allocations: 0 bytes)
  809.140 μs (6 allocations: 336 bytes)
  2.300 μs (0 allocations: 0 bytes)
  809.392 μs (6 allocations: 336 bytes)
  2.166 μs (0 allocations: 0 bytes)
  906.218 μs (0 allocations: 0 bytes)
  810.070 μs (6 allocations: 336 bytes)
  796.803 μs (6 allocations: 336 bytes)
  790.676 μs (6 allocations: 336 bytes)
  796.903 μs (6 allocations: 336 bytes)
  791.014 μs (6 allocations: 336 bytes)
  905.247 μs (0 allocations: 0 bytes)
  810.338 μs (6 allocations: 336 bytes)
  796.873 μs (6 allocations: 336 bytes)
  790.946 μs (6 allocations: 336 bytes)
  796.773 μs (6 allocations: 336 bytes)
  790.640 μs (6 allocations: 336 bytes)

# with this PR
julia> for transformA in (identity, adjoint, transpose), transformD in (identity, Adjoint, Transpose)
           @btime mul!($(copy(B)), $(transformA(A)), $(transformD(D)))
           @btime mul!($(copy(B)), $(transformD(D)), $(transformA(A)))
       end
  1.982 μs (2 allocations: 96 bytes)
  2.238 μs (0 allocations: 0 bytes)
  1.954 μs (0 allocations: 0 bytes)
  2.145 μs (0 allocations: 0 bytes)
  1.971 μs (0 allocations: 0 bytes)
  2.133 μs (0 allocations: 0 bytes)
  21.606 μs (2 allocations: 96 bytes)
  22.907 μs (0 allocations: 0 bytes)
  21.615 μs (3 allocations: 112 bytes)
  22.823 μs (1 allocation: 16 bytes)
  21.529 μs (2 allocations: 96 bytes)
  22.920 μs (0 allocations: 0 bytes)
  21.543 μs (2 allocations: 96 bytes)
  22.940 μs (0 allocations: 0 bytes)
  21.546 μs (3 allocations: 112 bytes)
  23.016 μs (1 allocation: 16 bytes)
  21.588 μs (2 allocations: 96 bytes)
  22.967 μs (0 allocations: 0 bytes)
```
I know I keep asking the same question in (almost) all of my PRs: this is a pure performance improvement, all methods (as per the code above) have been working correctly before, so this might be worth backporting?