@dkarrasch 2019-06-19
This should be of interest to a number of people, so let me ping @andreasnoack, @dmbates, @mcognetta, @chriscoey and @kshyatt to make quick progress.