@StefanKarpinski 2019-07-02
Can we get a review from some linalg peeps? @kshyatt, @andreasnoack?